requirements.txt
alembic==1.2.1
altgraph==0.16.1
aniso8601==8.0.0
appnope==0.1.0
astroid==2.2.5
attrs==19.1.0
autograd==1.3
autograd-gamma==0.4.1
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
botometer==1.4
Bottleneck==1.2.1
branca==0.3.1
bs4==0.0.1
BSON-NumPy==0.2.0.dev0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorlover==0.3.0
cufflinks==0.16
cycler==0.10.0
dateparser==0.7.2
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
et-xmlfile==1.0.1
feedparser==5.2.1
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
folium==0.10.0
future==0.17.1
html5lib==1.0.1
humanize==0.5.1
idna==2.8
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.14.1
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter-client==5.3.1
jupyter-core==4.5.0
kiwisolver==1.1.0
langdetect==1.0.7
lazy-object-proxy==1.4.1
lifelines==0.22.2
lineup-widget==1.0.7
localconfig==1.1.3
lxml==4.3.4
macholib==1.11
Mako==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
maya==0.6.1
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
more-itertools==7.2.0
mysql-connector-python==8.0.18
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.4
nltk==3.4.5
notebook==6.0.0
numpy==1.17.3
oauthlib==3.1.0
openpyxl==2.6.2
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.1
patsy==0.5.1
pendulum==2.0.5
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
plotly==3.0.0
preprocessor==1.1.3
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.10.0
ptyprocess==0.6.0
py-mysql==1.0
Pygments==2.4.2
PyInstaller==3.5
pylint==2.3.1
pymongo==3.9.0
PyMySQL==0.9.3
pyparsing==2.4.1
pyrsistent==0.15.3
PySocks==1.7.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
pytzdata==2019.3
pyzmq==18.0.2
regex==2019.8.19
requests==2.22.0
requests-oauthlib==1.2.0
retrying==1.3.3
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
snaptime==0.2.4
soupsieve==1.9.2
spotipy==2.4.4
SQLAlchemy==1.3.5
statsmodels==0.10.1
terminado==0.8.2
testpath==0.4.2
textblob==0.15.3
torch==1.3.0.post2
torchvision==0.4.1.post2
tornado==6.0.3
traitlets==4.3.2
tweepy==3.8.0
tweet-preprocessor==0.5.0
typed-ast==1.4.0
tzlocal==2.0.0
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.0
wordcloud==1.5.0
wrapt==1.11.2
xlrd==1.2.0
